+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## Changed
+## [0.3.1] - 16-03-2020
+
+### Changed
 
 - Changed `layout` for application logging to log `X-Request-ID`
 
@@ -100,6 +102,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
+[0.3.1]: https://github.com/GeminiWind/service-libraries/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/GeminiWind/service-libraries/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/GeminiWind/service-libraries/compare/0.1.2...0.2.0
 [0.1.2]: https://github.com/GeminiWind/service-libraries/compare/0.1.1...0.1.2